+++ oop/uim/oop/commands/command.d
@@ -17,5 +17,5 @@ class DCommand : UIMObject, ICommand {
     }
 
     // Implement this method with your command`s logic.
-    abstract ulong execute(Json[string] options, IConsole console = null);
+    abstract size_t execute(Json[string] options, IConsole console = null);
 }
